requirements.txt
openai>=1.40.0
requests>=2.32.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pydantic>=2.8.2
pydantic-settings>=2.4.0
